🚚 rename ::  entity > entities로 네이밍 변경
