+++ src/entities/comment/model.ts
