+++ src/entities/post/model.ts
