+++ src/entities/user/model.ts
